+++ tibiapy/builders/house.py
@@ -6,54 +6,53 @@ from typing import List, TYPE_CHECKING, Optional
 from tibiapy.models.house import HouseEntry, HousesSection, House
 
 if TYPE_CHECKING:
-    from tibiapy import HouseStatus, HouseType, HouseOrder
-
+    from tibiapy import HouseStatus, HouseType, HouseOrder, Sex
 
 
 class HousesSectionBuilder:
-    def __init__(self, **kwargs):
-        self._world = kwargs.get("world")
-        self._town = kwargs.get("town")
-        self._status = kwargs.get("status")
-        self._house_type = kwargs.get("house_type")
-        self._order = kwargs.get("order")
-        self._entries = kwargs.get("entries") or []
-        self._available_worlds = kwargs.get("available_worlds")
-        self._available_towns = kwargs.get("available_towns")
+    def __init__(self):
+        self._world = None
+        self._town = None
+        self._status = None
+        self._house_type = None
+        self._order = None
+        self._entries = []
+        self._available_worlds = []
+        self._available_towns = []
 
-    def world(self, world: str) -> 'HousesSectionBuilder':
+    def world(self, world: str):
         self._world = world
         return self
 
-    def town(self, town:  str) -> 'HousesSectionBuilder':
+    def town(self, town:  str):
         self._town = town
         return self
 
-    def status(self, status: HouseStatus) -> 'HousesSectionBuilder':
+    def status(self, status: HouseStatus):
         self._status = status
         return self
 
-    def house_type(self, house_type: HouseType) -> 'HousesSectionBuilder':
+    def house_type(self, house_type: HouseType):
         self._house_type = house_type
         return self
 
-    def order(self, order: HouseOrder) -> 'HousesSectionBuilder':
+    def order(self, order: HouseOrder):
         self._order = order
         return self
 
-    def entries(self, entries: List[HouseEntry]) -> 'HousesSectionBuilder':
+    def entries(self, entries: List[HouseEntry]):
         self._entries = entries
         return self
 
-    def add_entry(self, entry: HouseEntry) -> 'HousesSectionBuilder':
+    def add_entry(self, entry: HouseEntry):
         self._entries.append(entry)
         return self
 
-    def available_worlds(self, available_worlds: List[str]) -> 'HousesSectionBuilder':
+    def available_worlds(self, available_worlds: List[str]):
         self._available_worlds = available_worlds
         return self
 
-    def available_towns(self, available_towns: List[str]) -> 'HousesSectionBuilder':
+    def available_towns(self, available_towns: List[str]):
         self._available_towns = available_towns
         return self
 
@@ -71,60 +70,60 @@ class HousesSectionBuilder:
 
 
 class _HouseWithIdBuilder:
-    def __init__(self, **kwargs):
-        self._id = kwargs.get("id")
-        self._name = kwargs.get("name")
-        self._world = kwargs.get("world")
+    def __init__(self):
+        self._id = None
+        self._name = None
+        self._world = None
 
-    def id(self, id: int) -> '_HouseWithIdBuilder':
+    def id(self, id: int):
         self._id = id
         return self
 
-    def world(self, world: int) -> '_HouseWithIdBuilder':
+    def world(self, world: str):
         self._world = world
         return self
 
-    def name(self, name: int) -> '_HouseWithIdBuilder':
+    def name(self, name: str):
         self._name = name
         return self
 
 
 class HouseEntryBuilder(_HouseWithIdBuilder):
-    def __init__(self, **kwargs):
-        super().__init__(**kwargs)
-        self._town = kwargs.get("town")
-        self._size = kwargs.get("size")
-        self._status = kwargs.get("status")
-        self._type = kwargs.get("type")
-        self._rent = kwargs.get("rent")
-        self._time_left = kwargs.get("time_left")
-        self._highest_bid = kwargs.get("highest_bid")
+    def __init__(self):
+        super().__init__()
+        self._town = None
+        self._size = None
+        self._status = None
+        self._type = None
+        self._rent = None
+        self._time_left = None
+        self._highest_bid = None
 
-    def status(self, status: HouseStatus) -> 'HouseBuilder':
+    def status(self, status: HouseStatus):
         self._status = status
         return self
 
-    def type(self, type: HouseType) -> 'HouseEntryBuilder':
+    def type(self, type: HouseType):
         self._type = type
         return self
 
-    def town(self, town: int) -> 'HouseEntryBuilder':
+    def town(self, town: str):
         self._town = town
         return self
 
-    def size(self, size: int) -> 'HouseEntryBuilder':
+    def size(self, size: int):
         self._size = size
         return self
 
-    def rent(self, rent: int) -> 'HouseEntryBuilder':
+    def rent(self, rent: int):
         self._rent = rent
         return self
 
-    def time_left(self, time_left: Optional[datetime.timedelta]) -> 'HouseEntryBuilder':
+    def time_left(self, time_left: Optional[datetime.timedelta]):
         self._time_left = time_left
         return self
 
-    def highest_bid(self, highest_bid: Optional[int]) -> 'HouseEntryBuilder':
+    def highest_bid(self, highest_bid: Optional[int]):
         self._highest_bid = highest_bid
         return self
 
@@ -142,86 +141,86 @@ class HouseEntryBuilder(_HouseWithIdBuilder):
 
 
 class HouseBuilder(_HouseWithIdBuilder):
-    def __init__(self, **kwargs):
-        super().__init__(**kwargs)
-        self._status = kwargs.get("status")
-        self._type = kwargs.get("type")
-        self._rent = kwargs.get("rent")
-        self._beds = kwargs.get("beds")
-        self._size = kwargs.get("size")
-        self._image_url = kwargs.get("image_url")
-        self._owner = kwargs.get("owner")
-        self._owner_sex = kwargs.get("owner_sex")
-        self._paid_until = kwargs.get("paid_until")
-        self._transfer_date = kwargs.get("transfer_date")
-        self._transferee = kwargs.get("transferee")
-        self._transfer_price = kwargs.get("transfer_price")
-        self._transfer_accepted = kwargs.get("transfer_accepted")
-        self._highest_bid = kwargs.get("highest_bid")
-        self._highest_bidder = kwargs.get("highest_bidder")
-        self._auction_end = kwargs.get("auction_end")
+    def __init__(self):
+        super().__init__()
+        self._status = None
+        self._type = None
+        self._rent = None
+        self._beds = None
+        self._size = None
+        self._image_url = None
+        self._owner = None
+        self._owner_sex = None
+        self._paid_until = None
+        self._transfer_date = None
+        self._transferee = None
+        self._transfer_price = None
+        self._transfer_accepted = None
+        self._highest_bid = None
+        self._highest_bidder = None
+        self._auction_end = None
 
-    def status(self, status: HouseStatus) -> 'HouseBuilder':
+    def status(self, status: HouseStatus):
         self._status = status
         return self
 
-    def rent(self, rent: int) -> 'HouseBuilder':
+    def rent(self, rent: int):
         self._rent = rent
         return self
 
-    def type(self, type: HouseType) -> 'HouseBuilder':
+    def type(self, type: HouseType):
         self._type = type
         return self
 
-    def image_url(self, image_url: str) -> 'HouseBuilder':
+    def image_url(self, image_url: str):
         self._image_url = image_url
         return self
 
-    def beds(self, beds: int) -> 'HouseBuilder':
+    def beds(self, beds: int):
         self._beds = beds
         return self
 
-    def size(self, size: int) -> 'HouseBuilder':
+    def size(self, size: int):
         self._size = size
         return self
 
-    def owner(self, owner: int) -> 'HouseBuilder':
+    def owner(self, owner: Optional[str]):
         self._owner = owner
         return self
 
-    def owner_sex(self, owner_sex: int) -> 'HouseBuilder':
+    def owner_sex(self, owner_sex: Sex):
         self._owner_sex = owner_sex
         return self
 
-    def paid_until(self, paid_until: int) -> 'HouseBuilder':
+    def paid_until(self, paid_until: Optional[datetime.datetime]):
         self._paid_until = paid_until
         return self
 
-    def transfer_date(self, transfer_date: int) -> 'HouseBuilder':
+    def transfer_date(self, transfer_date: Optional[datetime.datetime]):
         self._transfer_date = transfer_date
         return self
 
-    def transferee(self, transferee: int) -> 'HouseBuilder':
+    def transferee(self, transferee: Optional[str]):
         self._transferee = transferee
         return self
 
-    def transfer_price(self, transfer_price: int) -> 'HouseBuilder':
+    def transfer_price(self, transfer_price: Optional[int]):
         self._transfer_price = transfer_price
         return self
 
-    def transfer_accepted(self, transfer_accepted: int) -> 'HouseBuilder':
+    def transfer_accepted(self, transfer_accepted: Optional[bool]):
         self._transfer_accepted = transfer_accepted
         return self
 
-    def highest_bid(self, highest_bid: int) -> 'HouseBuilder':
+    def highest_bid(self, highest_bid: Optional[int]):
         self._highest_bid = highest_bid
         return self
 
-    def highest_bidder(self, highest_bidder: str) -> 'HouseBuilder':
+    def highest_bidder(self, highest_bidder: Optional[str]):
         self._highest_bidder = highest_bidder
         return self
 
-    def auction_end(self, auction_end: int) -> 'HouseBuilder':
+    def auction_end(self, auction_end: Optional[datetime.datetime]):
         self._auction_end = auction_end
         return self